requirements.txt
"""
Requirements file for Low-Light Image Enhancement Project

Install all dependencies:
    pip install -r requirements.txt
"""

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web application
streamlit>=1.28.0

# Utilities
numpy>=1.24.0
tqdm>=4.65.0
PyYAML>=6.0
tensorboard>=2.13.0

# Optional: For NIQE and advanced metrics
# pyiqa>=0.1.7
